techstore-backend/moderation-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
python-multipart==0.0.12
